+++ packages/odata-service-inquirer/src/types.ts
@@ -2,9 +2,9 @@ import type { Annotations, ServiceProvider } from '@sap-ux/axios-extension';
 import type { Destination } from '@sap-ux/btp-utils';
 import type { CommonPromptOptions, YUIQuestion } from '@sap-ux/inquirer-common';
 import type { OdataVersion } from '@sap-ux/odata-service-writer';
-import type { CdsVersionInfo } from '@sap-ux/project-access';
 import type { BackendSystem } from '@sap-ux/store';
 import type { ListChoiceOptions } from 'inquirer';
+import type { CapService } from '@sap-ux/cap-config-writer';
 import type { EntityAnswer, NavigationEntityAnswer } from './prompts/edmx/entity-helper';
 import type { TableSelectionMode, TableType } from '@sap-ux/fiori-elements-writer';
 
@@ -177,8 +177,8 @@ export interface EntitySelectionAnswers {
 }
 
 export interface TableConfigAnswers {
-    [EntityPromptNames.tableType]: TableType;
-    [EntityPromptNames.hierarchyQualifier]: string;
+    [EntityPromptNames.tableType]?: TableType;
+    [EntityPromptNames.hierarchyQualifier]?: string;
 }
 
 export interface AnnotationGenerationAnswers {
@@ -187,11 +187,11 @@ export interface AnnotationGenerationAnswers {
 }
 
 export interface AlpTableConfigAnswers {
-    [EntityPromptNames.tableAutoHide]: boolean;
-    [EntityPromptNames.tableMultiSelect]: boolean;
-    [EntityPromptNames.tableSelectionMode]: TableSelectionMode;
-    [EntityPromptNames.presentationQualifier]: string;
-    [EntityPromptNames.smartVariantManagement]: boolean;
+    [EntityPromptNames.tableAutoHide]?: boolean;
+    [EntityPromptNames.tableMultiSelect]?: boolean;
+    [EntityPromptNames.tableSelectionMode]?: TableSelectionMode;
+    [EntityPromptNames.presentationQualifier]?: string;
+    [EntityPromptNames.smartVariantManagement]?: boolean;
 }
 
 /**
@@ -201,40 +201,6 @@ export type EntityRelatedAnswers = EntitySelectionAnswers &
     TableConfigAnswers &
     AnnotationGenerationAnswers &
     AlpTableConfigAnswers;
-
-export type CapRuntime = 'Node.js' | 'Java';
-
-export interface CapService {
-    /**
-     * The path to the CAP project.
-     */
-    projectPath: string;
-    /**
-     * The CDS info for the Cds instance that was used to compile the project when determining the service.
-     */
-    cdsVersionInfo?: CdsVersionInfo;
-    /**
-     * The name of the CAP service as identified by the cds model.
-     */
-    serviceName: string;
-    /**
-     * The URL path to the service, as specfied in the manifest.json of generated apps
-     * This is also provided as `OdataServicePromptAnswers` property `servicePath`
-     */
-    urlPath?: string;
-    /**
-     * The relative path (from the `projectPath`) to the service cds file.
-     */
-    serviceCdsPath?: string;
-    /**
-     * The runtime of the Cds instance that was used to compile the project when determining the service.
-     */
-    capType?: CapRuntime;
-    /**
-     * The relative path (from the `projectPath`) to the app folder
-     */
-    appPath?: string;
-}
 
 export interface CapServiceChoice extends ListChoiceOptions {
     value: CapService;