+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.32.0
+
+### Minor Changes
+
+-   e6f42df: Update @sap-ux/annotation-converter, @sap-ux/vocabularies-types, @sap/ux-cds-compiler-facade dependencies
+
+### Patch Changes
+
+-   Updated dependencies [e6f42df]
+    -   @sap-ux/fiori-annotation-api@0.4.0
+
+## 0.31.27
+
+### Patch Changes
+
+-   Updated dependencies [93ef8c1]
+    -   @sap-ux/project-access@1.29.2
+    -   @sap-ux/fiori-annotation-api@0.3.8
+
+## 0.31.26
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.1
+-   @sap-ux/fiori-annotation-api@0.3.7
+
 ## 0.31.25
 
 ### Patch Changes