+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.3.68
+
+### Patch Changes
+
+-   Updated dependencies [93ef8c1]
+    -   @sap-ux/project-access@1.29.2
+    -   @sap-ux/inquirer-common@0.6.7
+
+## 0.3.67
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.1
+-   @sap-ux/inquirer-common@0.6.6
+
+## 0.3.66
+
+### Patch Changes
+
+-   Updated dependencies [080bda2]
+    -   @sap-ux/inquirer-common@0.6.5
+
 ## 0.3.65
 
 ### Patch Changes